🚚 move contribution aggregate_root in contribution module
