+++ marketplace-domain/src/contribution/aggregate.rs
